requirements.txt
# Core
python-telegram-bot>=21.0
python-dotenv>=1.0
APScheduler>=3.10
trafilatura>=1.12

# LLM - Single package for all models via Antigravity Tools proxy
openai>=1.0

# Database
# SQLite is built-in

# HTTP
httpx>=0.27

# Content extraction
youtube-transcript-api>=0.6
beautifulsoup4>=4.12
